Refactoring: created files simulation.* Moved some methods from formation_super.c to simulation.c
