+++ Code/webots/controllers/formation_super/pso_ocba.c
@@ -5,16 +5,16 @@
 // Algorithm for High-Dimensional Multi-Robot Learning" (Di Mario, Navarro, Martinoli; IEEE 2015) //
 //                                                                                                //
 //      Initialize particles                                                                      //
-//          for N_i iterations do                                                                 //
+//          for NB_ITERATIONS iterations do                                                       //
 //              for N_p particles do                                                              //
 //                  evaluate new particle position n0 times                                       //
 //                                                                                                //
-//              remaining_budget = iteration budget - n0 * N_p                                    //
+//              remaining_budget = iteration_budget - n0 * N_p                                    //
 //              while remaining_budget > 0                                                        //
 //                  Allocate DELTA samples among current positions and personal bests using OCBA. //
 //                  Evaluate allocated samples.                                                   //
 //                  Recalculate mean and variance for new evaluations.                            //
-//                  Remaining budget = remaining budget - DELTA                                   //
+//                  remaining_budget = remaining_budget - DELTA                                   //
 //                                                                                                //
 //              for N_p particles do                                                              //
 //                  Update personal best                                                          //
@@ -22,3 +22,9 @@
 //                  Update particle position                                                      //
 //                                                                                                //
 ////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#include <math.h>
+#include <stdio.h>
+#include <string.h>
+
+#include "pso_ocba.h"
